blah

+++ akatpp-x/akat_x_buzzer_freq.c
@@ -1,4 +1,4 @@
-static AKAT_UNUSED AKAT_PURE u8 ${oname}__get_prescaler(u16 const freq) {
+static AKAT_UNUSED AKAT_PURE u16 ${oname}__get_prescaler(u16 const freq) {
     % if len(splits) == 1:
     return ${splits[1]};
     % endif
